fix: updated package name, removed `core` fix: move unit tests to the root folder
