+++ package.json
@@ -1,5 +1,5 @@
 {
-    "name": "symbol-sdk-core-typescript",
+    "name": "symbol-sdk-typescript",
     "version": "0.0.1",
     "description": "symbol typescript sdk",
     "main": "dist/index.js",
@@ -33,19 +33,19 @@
         "@types/chai": "^4.2.21",
         "@types/jsonstream": "^0.8.30",
         "@types/long": "^4.0.1",
-        "@types/node": "^16.4.6",
         "@types/mocha": "^9.0.0",
+        "@types/node": "^16.4.6",
         "@types/ripemd160": "^2.0.0",
         "@typescript-eslint/eslint-plugin": "^4.28.5",
         "@typescript-eslint/parser": "^4.28.5",
         "chai": "^4.3.4",
         "coveralls": "^3.1.1",
-        "nyc": "^15.1.0",
         "eslint": "^7.31.0",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-prettier": "^3.4.0",
         "JSONStream": "^1.3.5",
         "mocha": "^9.0.3",
+        "nyc": "^15.1.0",
         "prettier": "^2.3.2",
         "prettier-plugin-organize-imports": "^2.3.3",
         "shx": "^0.3.3",
