+++ test/Key.spec.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 import { expect } from 'chai';
-import { Key } from '../../src/core/Key';
-import { Converter } from '../../src/core/utils';
+import { Key } from '../src/core/Key';
+import { Converter } from '../src/core/utils';
 
 describe('Key', () => {
     const testKey = Converter.hexToUint8('2E834140FD66CF87B254A693A2C7862C819217B676D3943267156625E816EC6F');
